Update kdeify for Debian packaging

+++ ci-tooling/lib/kdeify.rb
@@ -19,6 +19,11 @@ class KDEIfy
     end
 
     def install_kde_js
+      if Dir.exist?('debian/extra-stuff')
+        FileUtils.cp('../suse/MozillaFirefox/kde.js', 'debian/extra-stuff/')
+        return
+      end
+
       FileUtils.cp('../suse/MozillaFirefox/kde.js', 'debian/')
       rules = File.read('debian/rules')
       rules.gsub!(/pre-build.*$/) do |m|
@@ -30,9 +35,9 @@ class KDEIfy
     def add_plasma_package(package)
       # Add dummy package
       control = File.read('debian/control.in')
-      control += "\nPackage: @MOZ_PKG_NAME@-plasma
+      control += "\nPackage: @browser@-plasma
 Architecture: any
-Depends: @MOZ_PKG_NAME@ (= ${binary:Version}), mozilla-kde-support
+Depends: @browser@ (= ${binary:Version}), mozilla-kde-support
 Description: #{package} package for integration with KDE
  Install this package if you'd like #{package} with Plasma integration
 "
